02_activate/22_solution/requirements.txt
kfp
google-cloud-aiplatform
protobuf==3.20.3
google-cloud-pipeline-components==1.0.39
kfp==1.8.19
google-cloud-aiplatform==1.22.0
scikit-learn